Regenerate MIPS catan, catanh long double ulps.
This patch regenerates MIPS catan and catanh ulps for long double with
fixed expected results for the tests of those functions.  ulps for
other types (which may see variation depending on whether glibc is
built for a processor with fused multiply-add support) are
deliberately not reduced.  ulps are not regenerated for powerpc-nofpu
as such regeneration does not result in any changes for long double.

	* sysdeps/mips/mips64/libm-test-ulps: Update catan and catanh ulps
	for long double with corrected test expectations.