Fix tst-timezone race (bug 14096).
I observed a failure of timezone/tst-timezone with the same symptoms
as reported in bug 14096.

I was unable to reproduce the failure on testing again.  However, from
inspection, the cause is as follows.  tst-timezone uses time zones
compiled from the northamerica file, but has no dependency on any such
zones, so with a parallel build it is possible that they do not get
compiled until after tst-timezone is run.  This patch adds a
dependency on the compiled America/New_York zone (the one used as a
makefile target to cause such zones to be compiled, rather than one
which is actually used in that test) to fix the race.

Tested for x86_64.

	[BZ #14096]
	* timezone/Makefile ($(objpfx)tst-timezone.out): Depend on
	America/New_York.